move set_ipython_cell to utils
