+++ nglview/utils.py
@@ -14,6 +14,38 @@ if PY3:
 else:
     string_types = basestring
 
+def _set_ipython_cell():
+    cm = """
+    var setIPythonLikeCell = function(){
+        //var cell = Jupyter.notebook.insert_cell_at_bottom();
+        var cell = Jupyter.notebook.get_selected_cell();
+        var $el = cell.element;
+        cell.set_text('');
+        $el.css({'background': '#e5f5f9'});
+
+        var handler = function(event) {
+            var selected_cell = Jupyter.notebook.get_selected_cell();
+            if (selected_cell.cell_id === cell.cell_id){
+                selected_cell.execute();
+                selected_cell.set_text('');
+            }
+            return false;
+        };
+
+        var action = {
+            help: 'run cell',
+            help_index: 'zz',
+            handler: handler
+        };
+
+        Jupyter.keyboard_manager.edit_shortcuts.add_shortcut('enter', action); 
+    };
+
+    setIPythonLikeCell()
+    """
+    from IPython.display import display, Javascript, HTML
+    display(Javascript(cm))
+
 def ngl_demo(width=400, height=400):
     """make a viewport, create stage object and populate NGL namespace
     """
